tstSemXRoads: warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@41001 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/testcase/tstRTSemXRoads.cpp
@@ -124,13 +124,13 @@ static void tstTraffic(unsigned cThreads, unsigned cSecs)
     int rc = VINF_SUCCESS;
     for (unsigned i = 0; i < cThreads && RT_SUCCESS(rc); i++)
     {
-        rc = RTThreadCreateF(&ahThreadsX[i], tstTrafficNSThread, (void *)i, 0, RTTHREADTYPE_DEFAULT, RTTHREADFLAGS_WAITABLE, "NS-%u", i);
+        rc = RTThreadCreateF(&ahThreadsX[i], tstTrafficNSThread, (void *)(uintptr_t)i, 0, RTTHREADTYPE_DEFAULT, RTTHREADFLAGS_WAITABLE, "NS-%u", i);
         RTTEST_CHECK_RC_OK(g_hTest, rc);
     }
 
     for (unsigned i = 0; i < cThreads && RT_SUCCESS(rc); i++)
     {
-        rc = RTThreadCreateF(&ahThreadsX[i], tstTrafficEWThread, (void *)i, 0, RTTHREADTYPE_DEFAULT, RTTHREADFLAGS_WAITABLE, "NS-%u", i);
+        rc = RTThreadCreateF(&ahThreadsX[i], tstTrafficEWThread, (void *)(uintptr_t)i, 0, RTTHREADTYPE_DEFAULT, RTTHREADFLAGS_WAITABLE, "NS-%u", i);
         RTTEST_CHECK_RC_OK(g_hTest, rc);
     }
